requirements.txt
psutil>=5.9.0
pillow>=9.0.0
pystray>=0.19.0
